* фиксы для дробей (int) на (float) * jbimage добавлен data-template * дефолтное значение селекта для новой почты * Параметр image выбор картинки * Валидация с админки * Стили

+++ packages/jbuniversal/jbuniversal/framework/controllers/basket.php
@@ -174,7 +174,7 @@ class BasketJBUniversalController extends JBUniversalController
     public function quantity()
     {
         // get request
-        $value = (int)$this->_jbrequest->get('value');
+        $value = (float)$this->_jbrequest->get('value');
         $key   = trim($this->_jbrequest->get('key'));
 
         $cart = JBCart::getInstance();
